requirements.txt
# Westpac RCA Demo - Python Dependencies
# Generated: February 5, 2026

# Core Framework
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2

# Elasticsearch & Kibana
elasticsearch==8.12.0
eland==8.12.0

# AI/ML Libraries
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.23
sentence-transformers==2.3.1
transformers==4.37.0

# LLM Providers (Optional)
anthropic==0.18.0
openai==1.12.0

# Data Generation
faker==22.6.0
mimesis==12.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
pyarrow==15.0.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# HTTP Clients
httpx==0.26.0
requests==2.31.0

# Document Generation
python-docx==1.1.0
reportlab==4.0.9
jinja2==3.1.3

# Utilities
colorama==0.4.6
rich==13.7.0
tqdm==4.66.1
click==8.1.7

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.11.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
ipython==8.20.0
jupyter==1.0.0

# Jupyter Notebooks
notebook==7.0.7
ipywidgets==8.1.1
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Time & Date
pytz==2024.1
arrow==1.3.0

# Validation
jsonschema==4.21.0
cerberus==1.3.5

# Async
asyncio==3.4.3
aiohttp==3.9.3
aiofiles==23.2.1

# Caching
redis==5.0.1
diskcache==5.6.3
